Add first version of fully cached implementation
